Refactor packets into new header
