+++ src/OpenLoco/Network/NetworkBase.cpp
@@ -1,6 +1,7 @@
 #include "NetworkBase.h"
 #include "../Console.h"
 #include "../Platform/Platform.h"
+#include "Packet.h"
 
 using namespace OpenLoco;
 using namespace OpenLoco::Network;
